+++ src/hwacc/IOAcc.py
@@ -56,4 +56,4 @@ class IOAcc(BasicPioDevice):
     )
     gic = Param.BaseGic(Parent.any, "Gic on which to trigger interrupts")
     int_num = Param.UInt32(320, "Interrupt number that connects to GIC")
-    clock_period = Param.Int(10, "Clock period in ns")
+    clock_period = Param.Float(10, "Clock period in ns")
