+++ src/hwacc/LLVMInterface.py
@@ -46,12 +46,12 @@ class LLVMInterface(AccComputeUnit):
         "TRUE: Stall datapath if any operation stalls."
         "FALSE: Only stall datapath regions with stalls",
     )
-    sched_threshold = Param.UInt32(
+    sched_threshold = Param.Float(
         10000,
         "Scheduling window threshold. Prevents scheduling windows size"
         " from exploding during regions of high loop parallelism",
     )
-    clock_period = Param.Int32(10, "System clock speed")
+    clock_period = Param.Float(10, "System clock speed")
     top_name = Param.String(
         "top", "Name of the top-level function for the accelerator"
     )
