+++ src/hwacc/noncoherent_dma.hh
@@ -68,7 +68,7 @@ class NoncoherentDma : public DmaDevice
     unsigned maxReqSize;
     BaseGic * gic;
     uint32_t intNum;
-    int clock_period;
+    float clock_period;
 
     uint8_t * mmreg;
     uint8_t * FLAGS;
